rag_bot/requirements.txt
google-generativeai==0.8.3
langgraph==0.2.14
sentence-transformers==3.1.1
numpy==1.26.4
markdown==3.7
faiss-cpu==1.9.0